Refactor: Use one common router for OTP1 and 2
