+++ src/routes/v1/index.ts
@@ -1,0 +1,15 @@
+import { Router } from 'express'
+
+import nonTransitRouter from './non-transit'
+import transitRouter from './transit'
+import tripPatternsRouter from './trip-patterns'
+
+const router = Router()
+
+router.use('/transit', transitRouter)
+
+router.use('/trip-patterns', tripPatternsRouter)
+
+router.use('/non-transit', nonTransitRouter)
+
+export default router
