+++ src/routes/v1/otp2Areas.prod.ts
