+++ src/routes/v1/otp2Areas.staging.ts
